### 2025-08-01 #8695 @editorialbot
[PRE REVIEW]: MDPax: GPU-accelerated MDP solvers in Python with JAX
**Submitting author:** @joefarrington (<a href="http://orcid.org/0000-0003-4156-3419">Joseph Farrington</a>)
**Repository:** https://github.com/joefarrington/mdpax
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v0.2.0
