### 2025-08-25 #8695 @thchang
Hi @diehlpk I would also be happy to provide a review for this one.

Also my first time reviewing for JOSS, and it's been several years since my own JOSS paper was reviewed and I've somewhat forgotten the process.  If relevant, please also let me know of any relevant resources other than the reviewer guidelines already linked above.

In particular, I am wondering what is the recommended method for leaving any review comments that I might have:

 1. leave a single review comment on this thread?
 2. Open multiple issues on the `main` branch of the repository (if I have comments for the authors)?
 3. Some combination of these?
 4. Something else entirely?
