### 2025-09-14 #8897 @arrrunn
@diehlpk Here's my comments from the initial review, overall a good read but I haven't been able to install and check out important parts of the review yet to check functionality, examples etc. I'll list my comments in the order of the checklist -

General checks:
- Results: The paper claims through an earlier version used in their prior work solving the Hendrix problem in under 3.5 hours on consumer GPU. I'm not sure if this needs to be checked, but once installed, I can check. It is included in the repository as part of the problems.

Functionality:
- Installation: I had issues installing mdpax based on the instructions in the readme file. I'm attaching screenshots to show errors I faced, but it seems the main issue finally is with JAX. I see the other reviewer was able to install, so maybe I'm doing something wrong. Either way, more clear instructions to install under alternate dependencies that are not part of their project.toml would be helpful to verify functionality as well as claims of being able to use on CPU as well.
- Functionality: to be checked upon installation. Google collab - all except "Checkpointing" run without errors.
- Performance: Same as Results above in General Checks. The paper claims through an earlier version used in their prior work solving the Hendrix problem in under 3.5 hours on consumer GPU. I'm not sure if this needs to be checked, but once installed, I can check. It is included in the repository as part of the problems. Google collab results check out for the most part. Small inconsistencies exist as shown in the screenshot (under 10s when it took 14s for the initial problem).

Documentation:
- Installation instructions: for reasons mentioned above, it would help to provide more instructions around installation for the average user
-  Functionality documentation: The paper seems to have a very basic outline of the functionality. Examples included may support documentation but indirectly.
- Automated tests: Pending installation on CPU. Google collab - I was able to run the instance as provided to confirm all except "Checkpointing" example run as expected including performance guarantees for time for the most part. Small inconsistencies exist as shown in the screenshot (under 10s when it took 14s for the initial problem). I did initially run on 'CPU' on google collab and later updated to GPU instance to get close to performance expectations.

Software paper:
- References: TBD after checking installation + functionality.
