### 2025-09-14 #8897 @joefarrington
Hi @thchang and @arrrunn, thanks very much to you both for your time and this really helpful feedback.

I don't currently have access to a Mac so will have a look into what I can do about jax-metal and any other Mac dependency issues. I think everything else should be relatively straightforward so will start working on it this week and plan to spend some time next weekend on the testing strategy.
