### 2025-09-21 5f7ee5e
Update install instructions and dependecies for win, osx, colab
