+++ README.md
@@ -5,12 +5,12 @@
 [![Tests](https://github.com/joefarrington/mdpax/actions/workflows/tests.yml/badge.svg)](https://github.com/joefarrington/mdpax/actions/workflows/tests.yml)
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
 
-
 MDPax is designed for researchers and practitioners who want to solve large Markov Decision Process (MDP) problems but don't want to become experts in graphics processing unit (GPU) programming. By using JAX, we can take advantage of the massive parallel processing power of GPUs while describing new problems using a simple Python interface.
 
 You can run MDPax on your local GPU, or try it for free using [Google Colab](https://colab.research.google.com/), which provides access to GPUs in the cloud with no setup required.
 
 ## Key capabilities
+
 - Solve MDPs with millions of states using value iteration or policy iteration
 - Automatic support for one or more identical GPUs
 - Flexible interface for defining your own MDP problem or solver algorithm
@@ -19,15 +19,17 @@ You can run MDPax on your local GPU, or try it for free using [Google Colab](htt
 
 ## Overview
 
-MDPax is a Python package for solving large-scale MDPs, leveraging JAX's support for vectorization, parallelization, and just-in-time (JIT) compilation on GPUs. 
+MDPax is a Python package for solving large-scale MDPs, leveraging JAX's support for vectorization, parallelization, and just-in-time (JIT) compilation on GPUs.
 
 The package is adapted from the research code developed for [Farrington et al (2025)](https://doi.org/10.1007/s10479-025-06551-6) (a [preprint](https://arxiv.org/abs/2303.10672) was released in 2023). We demonstrated that this approach is particularly well-suited for perishable inventory management problems where the state space grows exponentially with the number of products and the maximum useful life of the products. By implementing the problems in JAX and using consumer-grade GPUs (or freely available GPUs on services such as Google Colab) it is possible to compute the exact solution for realistically sized perishable inventory problems where this was recently reported to be infeasible or impractical.
 
 Traditional value iteration implementations face two main challenges with large state spaces:
+
 1. Memory requirements - the full transition matrix grows with the square of the state space size
 2. Computational complexity - nested loops over states, actions, and possible next states become prohibitively expensive
 
 MDPax addresses these challenges by:
+
 1. Using a functional approach where users specify a deterministic transition function in terms of state, action, and random event, rather than providing the full transition matrix
 2. Leveraging JAX's transformations to optimize computation:
    - `vmap` to vectorize operations across states and actions
@@ -43,18 +45,43 @@ MDPax can be installed from PyPI using pip:
 ```bash
 pip install mdpax
 ```
+
+The main dependencies are:
+
+- jax
+- chex
+- numpyro
+- orbax
+- loguru
+- hydra-core
+- jaxtyping
+- numpy
+
+See [`pyproject.toml`](https://github.com/joefarrington/mdpax/blob/main/pyproject.toml) for the complete list of dependencies and version requirements.
+
+### GPU (recommended)
+
+MDPax is designed for GPU-accelerated computation and works best on Linux systems with NVIDIA GPUs.
+
+For GPU support, ensure your NVIDIA drivers and CUDA toolkit are compatible with JAX. See the [JAX installation guide](https://github.com/google/jax#installation) for details.
+
+### CPU only
+
+MDPax will automatically fall back to CPU on Linux if no GPU is detected, though performance will be significantly slower for large problems.
+
[…]
