### 2025-09-21 8d33141
Fix Windows env variable issue in tests action
