### 2025-09-21 408d411
Checkpoint interval causing windows test failure
