+++ tests/test_utils/test_checkpointing.py
@@ -27,12 +27,12 @@ def test_checkpoint_save_restore(tmp_path):
         gamma=0.99,
         epsilon=1e-4,
         checkpoint_dir=checkpoint_dir,
-        checkpoint_frequency=5,
+        checkpoint_frequency=1,
         verbose=0,
     )
 
     # Run for a few iterations
-    solver.solve(max_iterations=20)
+    solver.solve(max_iterations=5)
     values_before = solver.values.copy()
     iter_before = solver.iteration
 
@@ -63,12 +63,12 @@ def test_lightweight_checkpoint_save_load(tmp_path):
         gamma=0.99,
         epsilon=1e-4,
         checkpoint_dir=checkpoint_dir_first,
-        checkpoint_frequency=5,
+        checkpoint_frequency=1,
         verbose=0,
     )
 
     # Run for a few iterations
-    solver.solve(max_iterations=20)
+    solver.solve(max_iterations=5)
     values_before = solver.values.copy()
     iter_before = solver.iteration
 
