### 2025-09-21 1772ef7
Add delay to checkpointing tests for Windows IO
