+++ tests/test_utils/test_checkpointing.py
@@ -1,5 +1,7 @@
 """Tests for checkpointing functionality."""
 
+import time
+
 import jax.numpy as jnp
 import pytest
 
@@ -36,6 +38,9 @@ def test_checkpoint_save_restore(tmp_path):
     values_before = solver.values.copy()
     iter_before = solver.iteration
 
+    # Small delay to let async checkpointing finish
+    time.sleep(1)
+
     # Create new solver and load checkpoint
     new_solver = ValueIteration.restore(checkpoint_dir)
     # Check state matches
@@ -71,6 +76,9 @@ def test_lightweight_checkpoint_save_load(tmp_path):
     solver.solve(max_iterations=5)
     values_before = solver.values.copy()
     iter_before = solver.iteration
+
+    # Small delay to let async checkpointing finish
+    time.sleep(1)
 
     checkpoint_dir_second = tmp_path / "checkpoints" / "test_lightweight_save_load_2"
 
