+++ paper/paper.bib
@@ -1,345 +1,0 @@
-@book{sutton_reinforcement_2018,
-  address   = {Cambridge, MA, USA},
-  edition   = {2nd ed.},
-  title     = {Reinforcement {Learning}: {An} {Introduction}},
-  isbn      = {978-0-262-03924-6},
-  language  = {en},
-  publisher = {The MIT Press},
-  author    = {Sutton, Richard S. and Barto, Andrew G.},
-  year      = {2018}
-}
-
-@article{ortega_cuda_2019,
-  title    = {A {CUDA} approach to compute perishable inventory control policies using value iteration},
-  volume   = {75},
-  issn     = {1573-0484},
-  url      = {https://doi.org/10.1007/s11227-018-2692-z},
-  doi      = {10.1007/s11227-018-2692-z},
-  language = {en},
-  number   = {3},
-  urldate  = {2022-08-03},
-  journal  = {The Journal of Supercomputing},
-  author   = {Ortega, G. and Hendrix, Eligius MT and García, I.},
-  month    = mar,
-  year     = {2019},
-  pages    = {1580--1593}
-}
-
-@phdthesis{johannsson_gpu-based_2009,
-  address = {Reykjavík, Iceland},
-  type    = {M.Sc. dissertation},
-  title   = {{GPU}-based {Markov} decision process solver},
-  url     = {https://en.ru.is/media/skjol-td/MSThesis_ArsaellThorJohannsson.pdf},
-  urldate = {2023-02-15},
-  school  = {Reykjavík University},
-  author  = {Jóhannsson, Arsæll Pór},
-  year    = {2009}
-}
-
-@article{farrington_going_2025,
-  title      = {Going faster to see further: graphics processing unit-accelerated value iteration and simulation for perishable inventory control using {JAX}},
-  issn       = {1572-9338},
-  shorttitle = {Going faster to see further},
-  url        = {https://doi.org/10.1007/s10479-025-06551-6},
-  doi        = {10.1007/s10479-025-06551-6},
-  language   = {en},
-  urldate    = {2025-05-11},
-  journal    = {Annals of Operations Research},
-  author     = {Farrington, Joseph and Wong, Wai Keong and Li, Kezhi and Utley, Martin},
-  month      = mar,
-  year       = {2025},
-  keywords   = {Dynamic programming, Inventory, Markov decision processes, Reinforcement learning, Simulation}
-}
-
-@article{andersen_mdpsolver_2025,
-  title      = {{MDPSolver}: {An} {Efficient} {Solver} for {Markov} {Decision} {Processes}},
-  volume     = {10},
-  issn       = {2475-9066},
-  shorttitle = {{MDPSolver}},
-  url        = {https://joss.theoj.org/papers/10.21105/joss.07544},
-  doi        = {10.21105/joss.07544},
-  language   = {en},
-  number     = {109},
-  urldate    = {2025-07-23},
-  journal    = {Journal of Open Source Software},
-  author     = {Andersen, Anders Reenberg and Andersen, Jesper Fink},
-  month      = may,
-  year       = {2025},
-  pages      = {7544}
-}
-
-@article{gargiani_madupite_2025,
-  title      = {madupite: {A} {High}-{Performance} {Distributed} {Solver} for {Large}-{Scale} {Markov} {Decision} {Processes}},
-  volume     = {10},
-  issn       = {2475-9066},
-  shorttitle = {madupite},
-  url        = {https://joss.theoj.org/papers/10.21105/joss.07411},
-  doi        = {10.21105/joss.07411},
-  language   = {en},
-  number     = {108},
[…]
