+++ paper/paper.md
@@ -1,77 +1,0 @@
----
-title: "MDPax: GPU-accelerated MDP solvers in Python with JAX "
-tags:
-  - Python
-  - optimization
-  - dynamic programming
-  - reinforcement learning
-authors:
-  - name: Joseph Farrington
-    orcid: 0000-0003-4156-3419
-    affiliation: 1
-    corresponding: True
-  - name: Wai Keong Wong
-    affiliation: "1, 2, 3"
-  - name: Kezhi Li
-    affiliation: 1
-  - name: Martin Utley
-    affiliation: 4
-affiliations:
-  - name: Institute of Health Informatics, University College London, United Kingdom
-    index: 1
-  - name: NIHR University College London Hospitals Biomedical Research Centre, United Kingdom
-    index: 2
-  - name: Cambridge University Hospitals NHS Foundation Trust, United Kingdom
-    index: 3
-  - name: Clinical Operational Research Unit, University College London, United Kingdom
-    index: 4
-
-date: 1 August 2025
-bibliography: paper.bib
----
-
-# Summary
-
-MDPax is a Python library for solving large-scale Markov decision processes (MDPs), leveraging JAX’s [@bradbury_jax_2022] support for vectorization, parallelization, and just-in-time (JIT) compilation on graphics processing units (GPUs). It includes GPU-accelerated implementations of standard algorithms including value iteration and policy iteration [@sutton_reinforcement_2018].
-
-MDPs describe sequential decision-making problems in which, at each timestep, an agent observes the current state of its environment, selects an action, transitions to a new state by taking the selected action, and receives a reward. The goal is to find a policy (a mapping from observed states to actions) that maximizes the expected long-term reward, accounting for both the immediate and future consequences of actions. MDPs have been used to model a wide range of problems, including medical treatment planning [@schaefer_modeling_2004], traffic light control [@haijema_dynamic_2017], financial portfolio management [@bauerle_markov_2011], and conservation policy [@nicol_conservation_2010].
-
-Exact solution methods based on dynamic programming scale poorly due to the curse of dimensionality [@bellman_dynamic_1957]: the number of states can grow exponentially with problem parameters, quickly making computation very challenging. Deep reinforcement learning has proven to be very effective at finding approximate solutions for large problems [@arulkumaran_deep_2017], but exact solutions remain valuable both in their own right and for benchmarking approximate methods.
-
-The exact algorithms are well suited to parallel execution, and modern GPUs have thousands of processing cores over which updates can be effectively distributed. By building on JAX, MDPax makes it easy to take advantage of this hardware through a high-level Python API, enabling researchers and practitioners to solve problems with millions of states without needing expertise in GPU programming.
-
-# Statement of need
-
-MDPax was originally developed to support our work on perishable inventory management. Solving these problems exactly requires accounting not just for total inventory levels, but also for the age profile of the stock. As a result, the state space grows exponentially with the product’s maximum useful life, making realistic problem instances extremely large. Although many MDP solvers exist, exact methods are widely considered impractical or infeasible for realistically sized perishable inventory problems [@nahmias_perishable_1982; @hendrix_computing_2019; @de_moor_reward_2022; @abouee-mehrizi_platelet_2025].
-
-Implementations of exact solution methods face two main challenges when applied to large MDPs:
-
-- Memory requirements: The full transition matrix describing the dynamics of the problem grows quadratically with the number of states and linearly with the number of actions, making it infeasible to store and process for large problems.
-
-- Computational complexity: The core operations involve nested loops over states, actions, and successor states, which become prohibitively expensive as the state space grows.
-
[…]
