+++ .github/workflows/publish_to_pypi.yml
@@ -1,54 +1,0 @@
-name: Publish Python distribution to PyPI
-
-on:
-  release:
-    types: [created]
-
-jobs:
-  build:
-    name: Build distribution
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v4
-      with:
-        persist-credentials: false
-    - name: Set up Python
-      uses: actions/setup-python@v5
-      with:
-        python-version: "3.x"
-    - name: Install pypa/build
-      run: >-
-        python3 -m
-        pip install
-        build
-        --user
-    - name: Build a binary wheel and a source tarball
-      run: python3 -m build
-    - name: Store the distribution packages
-      uses: actions/upload-artifact@v4
-      with:
-        name: python-package-distributions
-        path: dist/
-
-  publish-to-pypi:
-      name: Publish Python distribution to PyPI
-      needs:
-      - build
-      runs-on: ubuntu-latest
-
-      environment:
-        name: pypi
-        url: https://pypi.org/p/mdpax
-
-      permissions:
-        id-token: write
-
-      steps:
-      - name: Download all the dists
-        uses: actions/download-artifact@v4
-        with:
-          name: python-package-distributions
-          path: dist/
-      - name: Publish distribution to PyPI
-        uses: pypa/gh-action-pypi-publish@release/v1
