### 2025-09-21 #8897 @joefarrington
Hi @thchang, thanks again for the helpful feedback. I've responded to the required/optional points below. I still need to do some work restructuring/adding tests but think I have covered everything else. I've published a new version of the package to PyPI (v0.2.1) incorporating these changes.

## Installation
>
>     * Required: Currently seems to only support Linux with NVIDIA GPUs
>       (jax[cuda]) -- this should be clarified somewhere in the README and docs
>       along with any other system requirements

Updated installation instructions in the README to clarify that GPU support only for Linux at the moment (due to JAX). Updated pyproject.toml to use CPU JAX if installed on Windows or macOS (and added these platforms to tests in GitHub actions).

>     * Required: The dependencies can only be found in the `pyproject.toml`
>       file. To satisfy the JOSS requirement, please also list them in the
>       documentation somewhere.

Added a list of dependencies to installation section of the README, along with link to the pyproject.toml file.

>     * Optional: I would love to see an option to install and use `jax[cpu]` so
>       users can play with this on a local machine before moving to a GPU
>       machine (though this is not required for this review process).  I tried
>       it on my laptop and your code actually does appear to work out-of-the-box
>       on a macbook air with `jax[cpu]`; but the "slow" tests do fail with message
>       `Killed: 9` from the system.  The default install instructions didn't
>       work though because `jax[cuda]` does not exist on MacOS.

The jax[cuda] dependency will gracefully fall back to CPU-only if installed on a Linux machine without a GPU (albeit after potentially downloading some unnecessary dependencies). As above, have updated pyproject.toml to use CPU version of JAX if installed on Windows or macOS.

>     * Optional: I would love to see support for jax-metal so I can run on my
>       macbook using MPS instead of CUDA cores.  Much of the time, this can
>       actually work out-of-the-box, but I got some errors for your code.  I
>       don't know how hard it would be to fix these issues, but if it is not
>       hard, I would highly recommend doing so.

I think this is a good idea but I don't currently have access to an M-series Mac to test this and fix the bugs.

 ## Tests
>
>     * Required: The default tests take too long and require too many compute
>       resources to reasonably verify an installation.  You should better
>       document the `-m "no slow"` option for `pytest`, which I only found by
>       reading your `pyproject.toml`, but did not see in the documentation
>       anywhere.  A user should not need high-end GPUs to verify the
>       installation.  E.g., when installing on a HPC where the login nodes are
>       not configured with GPUs, and it is pretty inconvenient to submit a job
>       just to verify the installation.

Added documentation of the "no slow" command as an example in the Development section of the README, highlighting that these tests are suitable for running on CPU only.

>     * Required: There is no distinction between unit and regression/integration
>       tests for this project. As far as I can tell, the majority of the solver
>       tests are actually regression tests.  I don't see any unit tests for the
>       solve function or really any of the code in the either of the ABCs in
>       `src/mdpax/core`.  I would recommend creating a unit test for the problem
>       and solver ABCs, creating a test implementation of each that does
>       something very simple and verifiable (e.g., imagine a constant state
>       transition function for the problem and a deterministic greedy solver).
>       Then in the unit tests, just assert that these very simple state
>       transitions and solver steps behave as expected one-by-one when fed into
>       the existing code in your solver and problem ABCs.  I think some of this
>       is already done in the problem and solver tests, but those are really
>       mainly testing the implementations not the ABCs.  I would also highly
>       recommend that you separate unit tests (test an individual function and
>       nothing else) from regression tests (solve a problem).  Then add an
>       option to your `pyproject.toml` to run just unit tests, just regression
>       tests, or both.

Still TODO, but I have started thinking the best way to do this. My general approach was using the simple forest problem from pymdptoolbox to check the solver/problem classes worked overall but agree it would be useful to have something for the ABCs themselves and group the tests by role.

>     * Required: There is no code coverage for your tests, so we have no idea
>       how much of the code is being tested. It shouldn't be too hard to add
>       this using `pytest-cov` package.

Added: 84% coverage for the "no slow" tests, and 88% when using all tests.

## Documentation

>     * Required: It was a little hard for me to find your full documentation and API
>       buried in the `README.md`.  Consider putting a link to your documentation
>       site (mdpax.readthedocs.io) in the project description somewhere so it is
>       immediately obvious to any potential users.

Added a link in project “About” section on the top left of the GitHub page. There is also a “docs” badge at the top of the README that links to the docs.

 ## Performance
>     * Required: Not in the paper, but in the documentation the authors make some
>       performance claims that are supposedly verifiable in the notebook via Google
>       colab. However, when I logged into Google colab and ran the notebook (with
>       no warnings or errors) I did not reproduce the performance reported in the
>       documentation. Case 1 took about 6 seconds for setup + 39 seconds to solve
>       (the documentation suggests both of these together should finish in less
>       than 10 seconds). Case 2 took about 15 mins to finishing one single
>       iteration before I killed it (the docs reported it would finish in 3 mins or
>       less on a Google colab GPU).  To me, this seems like I may not be running on
>       a GPU, but how is this possible since the code clearly only installed
>       `jax[cuda]` and therefore should not work without a GPU?  Please see if you
>       can reproduce this issue, and suggest what I might be doing wrong or adjust
>       the numbers to reflect the range of all possible values on Google colab GPUs.

I think this was running on a CPU instance. As above, on Linux installing jax[cuda] works without a GPU, falling back to CPU (and this worked for me today on a CPU-based Colab instance). There are instructions in the README about how to check Colab is using a GPU instance. Rerunning Colab this week with GPU a few times, I found that (as arrrunn reported) Case 1 was often slower than 10s. I have changed this to 20s to provide some margin. I found that Case 2 consistently ran in less than 3 mins as originally claimed.
