### 2025-09-21 #8897 @joefarrington
Hi @arrrunn, thanks again for your help with this. I've responded to the points I thought needed a response at this stage below, and I've published a new version of the package to PyPI (v0.2.1) incorporating these changes.

## Functionality:

>     * Installation: I had issues installing mdpax based on the instructions in the readme file. I'm attaching screenshots to show errors I faced, but it seems the main issue finally is with JAX. I see the other reviewer was able to install, so maybe I'm doing something wrong. Either way, more clear instructions to install under alternate dependencies that are not part of their project.toml would be helpful to verify functionality as well as claims of being able to use on CPU as well.

I think this is a JAX/macOS issue. I've updated the installation instructions in the README, updated pyproject.toml to use CPU JAX if installed on Windows or macOS (and added these platforms for testing in GitHub actions). Tests pass for the macOS platform in GitHub actions so hopefully it will work for you too.

>     * Functionality: to be checked upon installation. Google collab - all except "Checkpointing" run without errors.

The issue with checkpointing was due to conflict between updates to orbax (used for checkpointing) and version of JAX supported by Colab. Fixed by pinning a version of orbax-checkpoint in colab-examples section of pyproject.toml. This may need to be updated when Google updates the version of JAX supported by Colab.

>     * Performance: Same as Results above in General Checks. The paper claims through an earlier version used in their prior work solving the Hendrix problem in under 3.5 hours on consumer GPU. I'm not sure if this needs to be checked, but once installed, I can check. It is included in the repository as part of the problems. Google collab results check out for the most part. Small inconsistencies exist as shown in the screenshot (under 10s when it took 14s for the initial problem).
I've rerun a few times this week and found the same thing for Case 1 so have updated the claim from 10s -> 20s to give some leeway. I found Case 2 still consistently finished within the claimed 3 minutes.

## Documentation:
>     * Installation instructions: for reasons mentioned above, it would help to provide more instructions around installation for the average user

As above, updated installation instruction to be clearer about supported platforms and CPU-fallback.

>     * Automated tests: Pending installation on CPU. Google collab - I was able to run the instance as provided to confirm all except "Checkpointing" example run as expected including performance guarantees for time for the most part. Small inconsistencies exist as shown in the screenshot (under 10s when it took 14s for the initial problem). I did initially run on 'CPU' on google collab and later updated to GPU instance to get close to performance expectations.

As above, fixed checkpointing dependency conflict for Colab and updated claim from 10s->20s in the notebook for Case 1.
