### 2025-09-22 ffc7e3b
Bump version from 0.2.0 to 0.2.1
