### 2025-10-05 #8897 @arrrunn
Thanks @diehlpk. I'm checking through all the comments now and I'll have an update this week. I'll post if I have any questions regarding handling specific resolutions if they're allowed to be accepted as is.
