### 2025-10-06 #8897 @thchang
Thanks @joefarrington ,

This addresses most of my issues.  I would still like to see unit tests for the solver ABC methods, as discussed, which is a larger challenge.  But based on the coverage report, I think the coverage is pretty good in those files so it is OK.

For the performance issue, I have re-run and was able to reproduce the result.  I think I did not do something right when loading the GPUs last time, I'm just surprised that it still ran since jax normally raises an error when using `jax[cuda]` and no GPUs.  Anyway, I followed the instructions again to activate the GPU and it looks like it is reproducing your runtimes now (with the update to the smaller problem).

All of my other concerns were pretty clearly addressed.  For me, this is sufficient that I have checked off all the boxes on my review checklist.

I have two final extremely minor nitpicks for the README file:
 1. When running with `jax[cuda]` on a linux machine with no GPUs (as I did this time around), you need to set the following environment variable:  `export JAX_PLATFORMS=cpu`.  Please add this to the documentation under the CPU section.  I know the focus is on GPUs, but making it easy so users can confirm installation on CPU only nodes (even if they will run on GPU nodes) really helps with usability of the package.
 2. It was not obvious to me how to generate the coverage report.  Normally, I expect the `pyproject.toml` to be configured such that the coverage is always recorded, and you only need to run `coverage report` afterward to view the stats.  I'm not totally sure, but that didn't appear to be the case for me.  I had to add the `--cov=src` option to get the coverage stats.  Whatever is your recommendation on how to generate the coverage, please just add an additional line to the Development subsection mentioning what options/extra commands are needed to get the coverage report.  It will be important for any other contributors besides yourself.

@diehlpk  I have finished my review and checked off all the boxes besides these 2 small things (above), I have no other comments on this paper, and to me it is ready for acceptance as soon as these minor documentation issues are fixed.

I saw the other reviewer still has a performance concern, I will let them resolve that separately.  For what it's worth, my 2 cents on the matter:
 - It is a tricky thing to balance since the Colab notebook in question is really more a part of the documentation, and there are downsides to bogging down the documentation with too much statistical machinery, as opposed to it remaining a succinct set of clear instructions on how to use the package and verify the results are as expected.  That said, since this is a published work, it is also important to make sure the performance values are not misrepresented due to small sample size.  I don't really know what is the right answer here, but I'm just voicing this contradiction after reading the above discussion.  Perhaps keep it simple in-line, with a footnote where you explain the result in the context of a more statistically rigorous experiment (e.g., this number is based on taking the mean/max over W repetitions on X Colab GPUs of Y version on Z date)?

One other small comment, I was never able to test GPU usage outside of the Colab environment (which does not reflect the setup and install instructions) as I have no access to GPU machines, so the intended installation process remains unverified by me, but I am pretty confident that it is working since that is how @diehlpk and his collaborators likely install it and also the Google Colab notebook works as expected with GPUs.  It would be best if someone else with GPU access could confirm the installation process and large tests though.

Let me know if/when anything else is needed from me regarding this review, now that I have completed my checklist.  Otherwise, I will wait until the other reviewer's discussion is resolved.

@joefarrington , pending the 2 final changes and resolving any remaining issues with the other reviewer, congratulations on your fine work here.
