### 2025-10-11 7215e6f
Produce coverage report when running tests
