+++ pyproject.toml
@@ -97,6 +97,7 @@ markers = [
     "slow: marks tests as slow (deselect with '-m \"not slow\"')",
     "serial",
 ]
+addopts = "--cov=mdpax --cov-report=term"
 
 [tool.coverage.run]
 source = ["src"]
