### 2025-10-16 #8897 @arrrunn
Hi @joefarrington, I was able to check the functionality as well as ran through the testing post all changes you've posted from the other reviewer. I did notice that I'd already installed separately "jax[cpu]" during all the installation issues from earlier but I see the other reviewer covered that detail to default to it and you've addressed it as well. Overall, all my checkboxes are ticked now.

Regarding the performance reporting, the idea is to just ensure correct representation and not mislead or under/over report when other users would like to use or develop further from this. For example, in the original Hendrix paper, they call it a 'rough estimate of computational time', so I guess to @thchang's point, as long as it is not an outlier to see such performance gains as reported by you, it should be fine to just footnote specifics for the claims just as an FYI. Also, I was talking about the specifics for the performance claims in the main text of the paper (not Collab) where you talk how much faster your problem computes the original Hendrix problem versus their claims. This could be as simple as "Using MDPax the same algorithm (always/most of the time/on average) converged in under 'x' hours when tested on 'consumer grade CPUs'/'4 data-center grade GPUs' (footnote with specifics and caveats)". Collab notebook, I understand, is mainly support and I don't expect too much detail there. Again, this is just being careful in a publication with claims contrasting other published work to ensure correctness as well as avoiding any counter claims of lack of validity under any scenario.

P.S. I tried running the large problem from Hendrix where it had 7M+ states on my CPU and my computer crashed after Iteration 4. But the one you claim for 3 hours with 1.16M states (80 hours for the original) actually ran in under 20 minutes and that's pretty impressive.
