### 2025-10-26 94e5560
Avoid double logging in solver subclasses
