+++ pyproject.toml
@@ -16,16 +16,16 @@ classifiers = [
 ]
 dependencies = [
     # Linux: CUDA-enabled JAX
-    "jax[cuda12]>=0.4.30; sys_platform == 'linux'",
+    "jax[cuda12]>=0.6.2; sys_platform == 'linux'",
     # Other platforms: CPU-only
-    "jax>=0.4.30; sys_platform != 'linux'",
+    "jax>=0.6.2; sys_platform != 'linux'",
     "chex>=0.1.88",
     "numpyro>=0.16.1",
     "orbax>=0.1.9",
     "loguru>=0.7.3",
     "hydra-core>=1.3.2",
     "jaxtyping>=0.2.36",
-    "numpy>=1.26.0",
+    "numpy>=1.26.0"
 ]
 
 [project.optional-dependencies]
