### 2025-10-26 83d643f
Limit to Python < 3.14 for due to MacOS build failure on GitHub actions
