### 2025-10-30 3d0b1ef
Update performance claim, orcids
