+++ paper/paper.bib
@@ -241,6 +241,7 @@
   title   = {{MDPtoolbox}: {Markov} {Decision} {Processes} {Toolbox}},
   url     = {https://cran.r-project.org/web/packages/MDPtoolbox/index.html},
   journal = {The Comprehensive R Archive Network (CRAN)},
+  doi     = {10.32614/cran.package.mdptoolbox},
   author  = {Chades, Iadine and Chapron, Guillaume and Cros, Marie-Josee and Garcia, Frederick and Sabbadin, Regis},
   year    = {2017}
 }
