### 2025-10-30 7e78a27
Bump version from 0.2.1 to 0.2.2
