+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "mdpax"
-version = "0.2.1"
+version = "0.2.2"
 description = "GPU-accelerated MDP solvers in Python with JAX"
 authors = [{name = "Joe Farrington", email = "farrington.joe@gmail.com"}]
 license = { text = "MIT" }
