### 2025-10-31 #8897 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
