### 2025-10-31 0894598
case changes in paper.bib
