+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @book{sutton_reinforcement_2018,
   address   = {Cambridge, MA, USA},
   edition   = {2nd ed.},
-  title     = {Reinforcement {Learning}: {An} {Introduction}},
+  title     = {Reinforcement Learning: An Introduction},
   isbn      = {978-0-262-03924-6},
   language  = {en},
   publisher = {The MIT Press},
@@ -52,7 +52,7 @@
 }
 
 @article{andersen_mdpsolver_2025,
-  title      = {{MDPSolver}: {An} {Efficient} {Solver} for {Markov} {Decision} {Processes}},
+  title      = {{MDPSolver}: An Efficient Solver for {M}arkov Decision Processes},
   volume     = {10},
   issn       = {2475-9066},
   shorttitle = {{MDPSolver}},
@@ -69,7 +69,7 @@
 }
 
 @article{gargiani_madupite_2025,
-  title      = {madupite: {A} {High}-{Performance} {Distributed} {Solver} for {Large}-{Scale} {Markov} {Decision} {Processes}},
+  title      = {madupite: A High-Performance Distributed Solver for Large-Scale {M}arkov Decision Processes},
   volume     = {10},
   issn       = {2475-9066},
   shorttitle = {madupite},
@@ -118,7 +118,7 @@
 
 @book{bellman_dynamic_1957,
   address   = {Princeton, NJ, USA},
-  title     = {Dynamic {Programming}},
+  title     = {Dynamic Programming},
   publisher = {Princeton University Press},
   author    = {Bellman, R.},
   year      = {1957},
@@ -176,7 +176,7 @@
 }
 
 @article{abouee-mehrizi_platelet_2025,
-  title   = {Platelet {Inventory} {Management} with {Approximate} {Dynamic} {Programming}},
+  title   = {Platelet Inventory Management with Approximate Dynamic Programming},
   issn    = {1091-9856},
   url     = {https://pubsonline.informs.org/doi/abs/10.1287/ijoc.2023.0245},
   doi     = {10.1287/ijoc.2023.0245},
@@ -205,7 +205,7 @@
 }
 
 @article{egorov_pomdpsjl_2017,
-  title      = {{POMDPs}.jl: {A} {Framework} for {Sequential} {Decision} {Making} under {Uncertainty}},
+  title      = {{POMDPs}.jl: A Framework for Sequential Decision Making under Uncertainty},
   volume     = {18},
   issn       = {1533-7928},
   shorttitle = {{POMDPs}.jl},
@@ -219,7 +219,7 @@
 }
 
 @misc{gargiani_inexact_2024,
-  title     = {Inexact {Policy} {Iteration} {Methods} for {Large}-{Scale} {Markov} {Decision} {Processes}},
+  title     = {Inexact Policy Iteration Methods for Large-Scale {M}arkov Decision Processes},
   url       = {http://arxiv.org/abs/2404.06136},
   doi       = {10.48550/arXiv.2404.06136},
   urldate   = {2025-07-23},
@@ -238,16 +238,16 @@
 }
 
 @misc{chades_mdptoolbox_2017,
-  title   = {{MDPtoolbox}: {Markov} {Decision} {Processes} {Toolbox}},
-  url     = {https://cran.r-project.org/web/packages/MDPtoolbox/index.html},
+  title   = {{MDPtoolbox}: {M}arkov Decision Processes Toolbox},
+  url     = {https://cran.r-project.org/web/packages/MDPtoolbox},
   journal = {The Comprehensive R Archive Network (CRAN)},
   doi     = {10.32614/cran.package.mdptoolbox},
[…]
