### 2025-10-31 #8897 @danielskatz
👋 @joefarrington - I've suggested a bunch of minor changes in https://github.com/joefarrington/mdpax/pull/7 - please merge this, or let me know what you disagree with, then we can continue to publication
